Added files via upload
This is example output of the bowtie2 wrapper given a maximum mismatch of 1. Notice that a colon is used not only to separate the read id from the list of taxids but also that every additional colon denotes another edit distance.

RunID: taxids with edit distance 0(comma delimited) : taxids with edit distance 1 : taxids with edit distance 2 etc.